fix for #246.
If the incoming request has app id specified both in the header AND
the JSON body, then check that both are the same.
Flag the request as invalid if they are not the same.

If the app id is not specified in the headers, then pick it up
from the JSON body.

Moreover, remove the app config keys (app id, master key, js key,
.net key etc.) from the request body unconditionally - ie. even if those
are specified in the header. This is so that if the proxy fronting the
app is not able these keys from the body.

Signed-off-by: Kunal Gangakhedkar <kunal.gangakhedkar@gmail.com>