added PromiseWithoutRejectedCallbackFail
I added a function that demonstrates issue #456. I'm finding that if I don't specify a rejected callback for the Parse.Config.get() promise's Then method, it never reaches my When method and the function times out. I found this worked in hosted parse but not in the open source version.

https://github.com/ParsePlatform/parse-server/issues/456