[hooks] allow hooks to be refreshed after set interval
Hooks configuration can only be processed on a ParseServer restart. This
is far from ideal. A simple configuration of an endpoint or a disabling
of a hook should not require a process of draining the parse server,
cycling it, and the bringing it back up. Simple configurations
changes should take effect more easily.