fix(middlewares): handleParseErrors incorrect log error message
On `handleParseErrors` middleware, if the error is instance of `ParseError`, the logger logs just `Parse error:` and does not include the actual error message.
This is bad for adapters, since we're forcing implementations to make use of `meta` data to gather useful information about the error thrown.
In this case, we could use `toString` of the own `ParseError` that would print something like `ParseError: 141 Invalid function: "hello"`.