Apply `npm install` to package-lock.json
It's not clear what happened in #6478, but it appears that it dropped
the `ws` dependency of the `parse` package when it was updating the
local version of the dependency. This means that running `npm install`
dirties the working tree, which is surprising!

A similar problem seems to have occurred in #6504, that time related to
the `commander` package.

I've filed an issue for the surprising behavior from greenkeeper:
https://github.com/greenkeeperio/greenkeeper-lockfile/issues/250