Add `bson-ext` as optional dependency
The `bson-ext` package builds a C++ addon that the Mongo driver can use
to more efficiently serialize/deserialize when talking to Mongo. When
this isn't loaded Mongo uses the `bson` package, which implements the
same functionality in JS, so we can classify this as an "optional"
dependency and not block a build if it is unavailable.