Fix workaround to make join query run
In e182744d this workaround was added so that if we are explicitly
querying on this join table it would run instead of bailing out early.

(This is necessary because the table we are querying is part of a Parse
Relation, not a Parse Class).

Unfortunately, the logic was _slightly_ off, preventing the query from
actually running. Oops!